+++ handler/openid/flow_refresh_token.go
@@ -57,6 +57,7 @@ func (c *OpenIDConnectRefreshHandler) HandleTokenEndpointRequest(ctx context.Con
 
 	// We need to reset the expires at value
 	sess.IDTokenClaims().ExpiresAt = time.Time{}
+	sess.IDTokenClaims().Nonce = ""
 	return nil
 }